requirements (1).txt
numpyp
pandas
scikit_learn
Flask
gunicorn
joblib
Werkzeug
